Icons and price setted
listos los iconos y precio
