+++ client/src/App.js
@@ -2,11 +2,17 @@ import React from 'react'
 
 import ProductCard from './components/ProductCards/ProductCards'
 
+import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles'
+
+import theme from './themeConfig.js';
+
 function App() {
   return (
-    <div>
-      <ProductCard></ProductCard>
-    </div>
+  	<ThemeProvider theme = {theme}>
+  	    <div>
+      		<ProductCard></ProductCard>
+    	</div>	
+  	</ThemeProvider>
   );
 }
